+++ docs/html/search/all_2.js
@@ -1,11 +1,12 @@
 var searchData=
 [
-  ['changelog_0',['Changelog',['../d4/d40/md_CHANGELOG.html',1,'']]],
-  ['changelog_2emd_1',['CHANGELOG.md',['../d3/d02/CHANGELOG_8md.html',1,'']]],
-  ['common_20usage_20patterns_2',['Common Usage Patterns',['../d1/d57/md_guides_2Getting_01Started.html#autotoc_md20',1,'']]],
-  ['complete_20example_3',['Complete Example',['../d1/d57/md_guides_2Getting_01Started.html#autotoc_md18',1,'']]],
-  ['content_4',['Content',['../d3/d69/classSimpleRest_1_1Api_1_1SimpleRestRequestBody.html#a8a4e28fefbb3c498603f9aef5fa2569d',1,'SimpleRest::Api::SimpleRestRequestBody']]],
-  ['contentlength_5',['ContentLength',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#acef052ba5d9ab1a2404373ddf2df130f',1,'SimpleRest::Api::SimpleRestRequest']]],
-  ['contenttype_6',['ContentType',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#acc1174e742762df8375238f662a7cfc7',1,'SimpleRest.Api.SimpleRestRequest.ContentType'],['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#a55389ee63100e7bda9c6877137429375',1,'SimpleRest.Api.SimpleRestResponse.ContentType']]],
-  ['contenttypes_7',['ContentTypes',['../d4/dfc/classSimpleRest_1_1Api_1_1SimpleRestContentTypeParser.html#ad67acdc05404685ee90edb230e553f9d',1,'SimpleRest.Api.SimpleRestContentTypeParser.ContentTypes'],['../da/d0d/interfaceSimpleRest_1_1Api_1_1ISimpleRestContentTypeParser.html#a8613717709e7d1fb2d57c08660553507',1,'SimpleRest.Api.ISimpleRestContentTypeParser.ContentTypes']]]
+  ['changed_0',['Changed',['../d4/d40/md_CHANGELOG.html#autotoc_md30',1,'']]],
+  ['changelog_1',['Changelog',['../d4/d40/md_CHANGELOG.html',1,'']]],
+  ['changelog_2emd_2',['CHANGELOG.md',['../d3/d02/CHANGELOG_8md.html',1,'']]],
+  ['common_20usage_20patterns_3',['Common Usage Patterns',['../d1/d57/md_guides_2Getting_01Started.html#autotoc_md20',1,'']]],
+  ['complete_20example_4',['Complete Example',['../d1/d57/md_guides_2Getting_01Started.html#autotoc_md18',1,'']]],
+  ['content_5',['Content',['../d3/d69/classSimpleRest_1_1Api_1_1SimpleRestRequestBody.html#a8a4e28fefbb3c498603f9aef5fa2569d',1,'SimpleRest::Api::SimpleRestRequestBody']]],
+  ['contentlength_6',['ContentLength',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#acef052ba5d9ab1a2404373ddf2df130f',1,'SimpleRest::Api::SimpleRestRequest']]],
+  ['contenttype_7',['ContentType',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#acc1174e742762df8375238f662a7cfc7',1,'SimpleRest.Api.SimpleRestRequest.ContentType'],['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#a55389ee63100e7bda9c6877137429375',1,'SimpleRest.Api.SimpleRestResponse.ContentType']]],
+  ['contenttypes_8',['ContentTypes',['../d4/dfc/classSimpleRest_1_1Api_1_1SimpleRestContentTypeParser.html#ad67acdc05404685ee90edb230e553f9d',1,'SimpleRest.Api.SimpleRestContentTypeParser.ContentTypes'],['../da/d0d/interfaceSimpleRest_1_1Api_1_1ISimpleRestContentTypeParser.html#a8613717709e7d1fb2d57c08660553507',1,'SimpleRest.Api.ISimpleRestContentTypeParser.ContentTypes']]]
 ];