+++ docs/html/search/all_3.js
@@ -4,5 +4,7 @@ var searchData=
   ['debug_1',['DEBUG',['../df/dba/namespaceSimpleRest_1_1Api.html#a213e4074609286d386cb9de95b93bac8adc30ec20708ef7b0f641ef78b7880a15',1,'SimpleRest::Api']]],
   ['delete_2',['DELETE',['../df/dba/namespaceSimpleRest_1_1Api.html#ada4eac4d8de31c19a9aa2abd91c5a909a32f68a60cef40faedbc6af20298c1a1e',1,'SimpleRest::Api']]],
   ['delete_3',['Delete',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#a9c2e0a6248ff7f3a9d85d6315e72346f',1,'SimpleRest::Api::SimpleRestApi']]],
-  ['documentation_4',['Documentation',['../index.html#autotoc_md7',1,'']]]
+  ['dispose_4',['Dispose',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#a20516048a808763837e18e864b2dd319',1,'SimpleRest::Api::SimpleRestApi']]],
+  ['disposed_5',['Disposed',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#a2d944b93e323b206968b147112771bb0',1,'SimpleRest::Api::SimpleRestApi']]],
+  ['documentation_6',['Documentation',['../index.html#autotoc_md7',1,'']]]
 ];