+++ docs/html/search/all_7.js
@@ -2,8 +2,9 @@ var searchData=
 [
   ['handling_0',['JSON Request/Response Handling',['../d1/d57/md_guides_2Getting_01Started.html#autotoc_md15',1,'']]],
   ['hascompleted_1',['HasCompleted',['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#ad70801b4a0e5d9815fe23449a745a19b',1,'SimpleRest::Api::SimpleRestResponse']]],
-  ['head_2',['HEAD',['../df/dba/namespaceSimpleRest_1_1Api.html#ada4eac4d8de31c19a9aa2abd91c5a909ae15e216fc1c639f787b1231ecdfa1bf8',1,'SimpleRest::Api']]],
-  ['head_3',['Head',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#a4f468b38d7bd34fdb0382a8cdeeecf11',1,'SimpleRest::Api::SimpleRestApi']]],
-  ['headers_4',['Headers',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a17a7f4fcc4e9b52d7a54d4a4a0d80fbe',1,'SimpleRest::Api::SimpleRestRequest']]],
-  ['high_5',['HIGH',['../df/dba/namespaceSimpleRest_1_1Api.html#a213e4074609286d386cb9de95b93bac8ab89de3b4b81c4facfac906edf29aec8c',1,'SimpleRest::Api']]]
+  ['hasstarted_2',['HasStarted',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#afdbf06e5b1a4c238b07d67aedd3f813d',1,'SimpleRest::Api::SimpleRestApi']]],
+  ['head_3',['HEAD',['../df/dba/namespaceSimpleRest_1_1Api.html#ada4eac4d8de31c19a9aa2abd91c5a909ae15e216fc1c639f787b1231ecdfa1bf8',1,'SimpleRest::Api']]],
+  ['head_4',['Head',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#a4f468b38d7bd34fdb0382a8cdeeecf11',1,'SimpleRest::Api::SimpleRestApi']]],
+  ['headers_5',['Headers',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a17a7f4fcc4e9b52d7a54d4a4a0d80fbe',1,'SimpleRest::Api::SimpleRestRequest']]],
+  ['high_6',['HIGH',['../df/dba/namespaceSimpleRest_1_1Api.html#a213e4074609286d386cb9de95b93bac8ab89de3b4b81c4facfac906edf29aec8c',1,'SimpleRest::Api']]]
 ];