+++ docs/html/search/properties_3.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['hascompleted_0',['HasCompleted',['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#ad70801b4a0e5d9815fe23449a745a19b',1,'SimpleRest::Api::SimpleRestResponse']]],
-  ['headers_1',['Headers',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a17a7f4fcc4e9b52d7a54d4a4a0d80fbe',1,'SimpleRest::Api::SimpleRestRequest']]]
+  ['endpoint_0',['Endpoint',['../da/d05/classSimpleRest_1_1Api_1_1SimpleRestMap.html#a01928470a90bf7d66f076f878bd75a22',1,'SimpleRest.Api.SimpleRestMap.Endpoint'],['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a887b159a1571cc3ab9d55c15e6d04448',1,'SimpleRest.Api.SimpleRestRequest.Endpoint']]]
 ];