+++ docs/html/search/properties_4.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['keys_0',['Keys',['../de/db5/classSimpleRest_1_1Api_1_1SimpleRestQuery.html#aa5cb03ba7d682b0b950d2b642b5cc89e',1,'SimpleRest.Api.SimpleRestQuery.Keys'],['../da/d93/interfaceSimpleRest_1_1Api_1_1ISimpleRestQuery.html#aaa96e62bf341143184ecf5a5b938c0ab',1,'SimpleRest.Api.ISimpleRestQuery.Keys']]]
+  ['hascompleted_0',['HasCompleted',['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#ad70801b4a0e5d9815fe23449a745a19b',1,'SimpleRest::Api::SimpleRestResponse']]],
+  ['hasstarted_1',['HasStarted',['../d7/d9c/classSimpleRest_1_1Api_1_1SimpleRestApi.html#afdbf06e5b1a4c238b07d67aedd3f813d',1,'SimpleRest::Api::SimpleRestApi']]],
+  ['headers_2',['Headers',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a17a7f4fcc4e9b52d7a54d4a4a0d80fbe',1,'SimpleRest::Api::SimpleRestRequest']]]
 ];