+++ docs/html/search/properties_6.js
@@ -1,6 +1,5 @@
 var searchData=
 [
-  ['params_0',['Params',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a2e3c9637b64eeeac18126b7850073993',1,'SimpleRest::Api::SimpleRestRequest']]],
-  ['path_1',['Path',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a67f4d4d0493aca00bbfc101430bbb9ea',1,'SimpleRest::Api::SimpleRestRequest']]],
-  ['pattern_2',['Pattern',['../da/d05/classSimpleRest_1_1Api_1_1SimpleRestMap.html#a569f4d8b95700c0006b63b19344764af',1,'SimpleRest::Api::SimpleRestMap']]]
+  ['method_0',['Method',['../da/d05/classSimpleRest_1_1Api_1_1SimpleRestMap.html#aa67549aec1dfbf8de9b9fcab93a45668',1,'SimpleRest.Api.SimpleRestMap.Method'],['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#aea3f4b64f3db73d6b46f2ae57cc4d89b',1,'SimpleRest.Api.SimpleRestRequest.Method']]],
+  ['middleware_1',['Middleware',['../da/d05/classSimpleRest_1_1Api_1_1SimpleRestMap.html#a83bb42d7aae71924adb8e388b6693d62',1,'SimpleRest::Api::SimpleRestMap']]]
 ];