+++ docs/html/search/properties_8.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['response_0',['Response',['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#a9801537d19f71dc18580f13a13c9f6c6',1,'SimpleRest::Api::SimpleRestResponse']]]
+  ['query_0',['Query',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a0ad66807922596fcefd9eff3a34df02b',1,'SimpleRest.Api.SimpleRestRequest.Query'],['../de/db5/classSimpleRest_1_1Api_1_1SimpleRestQuery.html#ac3efc083439da9d1e9ac55ac3b5d0456',1,'SimpleRest.Api.SimpleRestQuery.Query'],['../da/d93/interfaceSimpleRest_1_1Api_1_1ISimpleRestQuery.html#aa2fb6f3a778137e0fcad4b8aa4de668a',1,'SimpleRest.Api.ISimpleRestQuery.Query']]]
 ];