+++ docs/html/search/properties_a.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['this_5bstring_20key_5d_0',['this[string key]',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#aac2d17d521f06371113b2294545b996e',1,'SimpleRest.Api.SimpleRestRequest.this[string key]'],['../de/db5/classSimpleRest_1_1Api_1_1SimpleRestQuery.html#a6ee8a0a7e2f094fb9ab3020be8191cd2',1,'SimpleRest.Api.SimpleRestQuery.this[string key]'],['../da/d93/interfaceSimpleRest_1_1Api_1_1ISimpleRestQuery.html#a7cc82b1af37ec66623cff907aa035f4b',1,'SimpleRest.Api.ISimpleRestQuery.this[string key]']]]
+  ['statuscode_0',['StatusCode',['../d1/d42/classSimpleRest_1_1Api_1_1SimpleRestResponse.html#ad112d935fe8e36a33ece30261a1808e7',1,'SimpleRest::Api::SimpleRestResponse']]]
 ];