+++ docs/html/search/properties_c.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['values_0',['Values',['../de/db5/classSimpleRest_1_1Api_1_1SimpleRestQuery.html#a0bb3a17cf3b73c4db20577e3225d32ad',1,'SimpleRest.Api.SimpleRestQuery.Values'],['../da/d93/interfaceSimpleRest_1_1Api_1_1ISimpleRestQuery.html#a7ceecfc7a21068ae200658308255129a',1,'SimpleRest.Api.ISimpleRestQuery.Values']]]
+  ['useragent_0',['UserAgent',['../d3/d48/classSimpleRest_1_1Api_1_1SimpleRestRequest.html#a34c034e17d01e86438b7a238a8c21b61',1,'SimpleRest::Api::SimpleRestRequest']]]
 ];